chore(tests/wallet): moved tests to match src
