+++ test/types/Wallet/methods/fromHDPublicKey.js
@@ -1,9 +1,9 @@
 const Dashcore = require('@dashevo/dashcore-lib');
 const { expect } = require('chai');
-const Wallet = require('../../../src/types/Wallet/Wallet');
-const fromHDPublicKey = require('../../../src/types/Wallet/methods/fromHDExtPublicKey');
-const gatherSail = require('../../fixtures/gathersail');
-const { WALLET_TYPES } = require('../../../src/CONSTANTS');
+const Wallet = require('../../../../src/types/Wallet/Wallet');
+const fromHDPublicKey = require('../../../../src/types/Wallet/methods/fromHDExtPublicKey');
+const gatherSail = require('../../../fixtures/gathersail');
+const { WALLET_TYPES } = require('../../../../src/CONSTANTS');
 /**
  * Theses first set of data labeled gatherSail correspond to the following mnemonic:
  * gather sail face invite together focus waste barely excuse slide harbor hint
