+++ test/types/Wallet/methods/fromMnemonic.js
@@ -1,8 +1,8 @@
 const Dashcore = require('@dashevo/dashcore-lib');
 const { expect } = require('chai');
-const fromMnemonic = require('../../../src/types/Wallet/methods/fromMnemonic');
-const knifeFixture = require('../../fixtures/knifeeasily');
-const { WALLET_TYPES } = require('../../../src/CONSTANTS');
+const fromMnemonic = require('../../../../src/types/Wallet/methods/fromMnemonic');
+const knifeFixture = require('../../../fixtures/knifeeasily');
+const { WALLET_TYPES } = require('../../../../src/CONSTANTS');
 
 describe('Wallet - fromMnemonic', () => {
   it('should indicate missing data', () => {
